+++ tests/Pest.php
@@ -7,4 +7,4 @@
  +-------------------------------------------------------------------------+
 */
 
-require_once __DIR__ . '/bootstrap.php';
+require_once __DIR__ . '/bootstrap-unit.php';